# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Small Squaretable API
  description: |
    Enterprise-grade multi-tenant SaaS platform for AI chat and character management.

    ## Authentication

    Most endpoints require authentication via JWT Bearer tokens. Include the access token in the Authorization header:

    ```
    Authorization: Bearer <your_access_token>
    ```

    ## Error Handling

    All errors follow a standard format:

    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Human-readable error message",
        "details": {}
      },
      "meta": {
        "timestamp": "2026-02-04T12:00:00.000Z"
      }
    }
    ```

    ## Pagination

    List endpoints support pagination via query parameters:

    - `page`: Page number (default: 1)
    - `limit`: Items per page (default: 20, max: 100)
    - `sortBy`: Field to sort by
    - `sortOrder`: Sort order (asc/desc, default: desc)

    ## Caching

    Some endpoints use Redis caching. Check the `X-Cache` response header:
    - `HIT`: Response served from cache
    - `MISS`: Response served from database

  version: 0.1.0
  contact:
    name: Small Squaretable Team
    url: https://github.com/small-squaretable
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.small-squaretable.com
    description: Production server
...
paths:
...
  /api/v1/characters:
    post:
      tags:
        - Characters
      summary: Create character
      description: Creates a new character for the authenticated user
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCharacterRequest'
      responses:
        '201':
          description: Character created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Character'
        '400':
          $ref: '#/components/responses/BadRequest'

    get:
      tags:
        - Characters
      summary: List user's characters
      description: Returns paginated list of the user's private characters
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Character list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PaginatedCharacters'
  /api/v1/characters/marketplace:
    get:
      tags:
        - Characters
      summary: Browse marketplace
      description: Returns paginated list of public characters (authentication optional)
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Marketplace characters
          headers:
            X-Cache:
              description: Cache status (HIT or MISS)
              schema:
                type: string
                enum: [HIT, MISS]
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PaginatedCharacters'
  /api/v1/characters/search:
    get:
      tags:
        - Characters
      summary: Search characters
      description: Full-text search across public characters
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          description: Search query
          required: true
          schema:
            type: string
            minLength: 1
        - name: category
          in: query
          description: Filter by category
          schema:
            type: string
        - name: tags
          in: query
          description: Filter by tags (comma-separated)
          schema:
            type: string
            example: "helper,coding"
        - name: sort
          in: query
          description: Sort order
          schema:
            type: string
            enum: [popular, newest, rating, relevance]
            default: relevance
        - name: filter
          in: query
          description: Additional filters
          schema:
            type: string
            enum: [all, safe, nsfw]
            default: safe
        - name: isNsfw
          in: query
          description: Include NSFW content
          schema:
            type: boolean
            default: false
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Search results
          headers:
            X-Cache:
              description: Cache status
              schema:
                type: string
                enum: [HIT, MISS]
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SearchResults'
  /api/v1/characters/{characterId}:
    get:
      tags:
        - Characters
      summary: Get character by ID
      description: Returns character details
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CharacterId'
      responses:
        '200':
          description: Character details
          headers:
            X-Cache:
              description: Cache status
              schema:
                type: string
                enum: [HIT, MISS]
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Character'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Characters
      summary: Update character
      description: Updates an existing character
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CharacterId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCharacterRequest'
      responses:
        '200':
          description: Character updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Character'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Characters
      summary: Delete character
      description: Permanently deletes a character
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CharacterId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/chats:
    post:
      tags:
        - Chats
      summary: Create chat
      description: Creates a new chat with a character
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChatRequest'
      responses:
        '201':
          description: Chat created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Chat'

    get:
      tags:
        - Chats
      summary: List chats
      description: Returns paginated list of user's chats
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Chat list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PaginatedChats'
...
  /ws:
    get:
      tags:
        - WebSocket
      summary: WebSocket connection
      description: |
        Establishes a WebSocket connection for real-time chat.

        Connection URL: `ws://localhost:3000/ws?token=<access_token>`

        ## Message Types

        ### Client -> Server
        - `CONNECTED` - Sent by server on successful connection
        - `USER_MESSAGE` - Send a message to the chat
        - `JOIN_CHAT` - Join a chat room
        - `LEAVE_CHAT` - Leave a chat room
        - `TYPING_START` - User started typing
        - `TYPING_STOP` - User stopped typing
        - `PING` - Heartbeat ping

        ### Server -> Client
        - `CONNECTED` - Connection established
        - `USER_MESSAGE` - New user message
        - `ASSISTANT_MESSAGE_CHUNK` - Streaming AI response chunk
        - `ASSISTANT_MESSAGE_DONE` - AI response complete
        - `USER_TYPING` - Another user is typing
        - `ERROR` - Error occurred
        - `PONG` - Heartbeat response
      operationId: websocketConnection
      parameters:
        - name: token
          in: query
          required: true
          description: JWT access token
          schema:
            type: string
      responses:
        '101':
          description: WebSocket connection established
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token (7 day expiration)
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sortBy
      in: query
      description: Field to sort by
      schema:
        type: string
    SortOrderParam:
      name: sortOrder
      in: query
      description: Sort order
      schema:
        type: string
        enum: [asc, desc]
        default: desc
    CharacterId:
      name: characterId
      in: path
      required: true
      description: Character ID
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          nullable: true
        error:
          type: object
          nullable: true
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
...
    Character:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 2000
          nullable: true
        avatarUrl:
          type: string
          format: uri
          maxLength: 2000
          nullable: true
        cardData:
          type: object
          additionalProperties: true
          description: SillyTavern character card data
        tags:
          type: array
          items:
            type: string
          maxItems: 20
        category:
          type: string
          maxLength: 50
          nullable: true
        isPublic:
          type: boolean
        isNsfw:
          type: boolean
        downloadCount:
          type: integer
          default: 0
        avgRating:
          type: number
          nullable: true
        ratingCount:
          type: integer
          default: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateCharacterRequest:
      type: object
      required: [name, cardData]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 2000
        avatarUrl:
          type: string
          format: uri
          maxLength: 2000
        cardData:
          type: object
          description: SillyTavern character card data
          additionalProperties: true
        tags:
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 20
        category:
          type: string
          maxLength: 50
        isNsfw:
          type: boolean
          default: false
    UpdateCharacterRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 2000
        avatarUrl:
          type: string
          format: uri
          maxLength: 2000
        cardData:
          type: object
          additionalProperties: true
        tags:
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 20
        category:
          type: string
          maxLength: 50
        isNsfw:
          type: boolean
...
    PaginatedCharacters:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Character'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    SearchResults:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Character'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
        query:
          type: string
          description: Original search query

    # ============================================================
    # Rating Schemas
    # ============================================================
...
    Chat:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        characterId:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 500
        metadata:
          type: object
          additionalProperties: true
        messageCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        character:
          $ref: '#/components/schemas/Character'
    CreateChatRequest:
      type: object
      required: [characterId]
      properties:
        characterId:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 500
        metadata:
          type: object
          additionalProperties: true
...
    PaginatedChats:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Chat'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: VALIDATION_ERROR
                  message:
                    type: string
                  details:
                    type: object
              meta:
                type: object
                properties:
                  timestamp:
                    type: string
                    format: date-time
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: NOT_FOUND
                  message:
                    type: string
                    example: The requested resource was not found
              meta:
                type: object
                properties:
                  timestamp:
                    type: string
                    format: date-time
...